I activated the apache2 module "mod_rename" according to https://www.drupal.org/node/134439 (i.e. Section "Step 1 - Method B: apache2.conf") With this enabled, I activated the drupal module "clean url". This module makes drupal urls better.
Unfortunately, during the configuration of the apache2-module
"rename", drupal could not display the neato generated files. I saved
them in the drupal temporary folder. And apparently the configuration
of /etc/apache2/apache2.conf made the files there unaccessible.  So I
made a new folder ~/loopediatest/sites/default/tmp/ and made it the
official drupal tmp folder. The rest I left unchanged. So that the
pictures are still generated in the folder
~/loopediatest/sites/default/files/tmp/ , which is now in the "files"
folder and publicly accessible.